feat(communication)：将数据传输方式由共享内存改为本地回环，并将数据类型切换为 float16 以提升性能
- 重新启用本地回环的数据传输方式以模拟真实万兆网线传输，但是减去了同步的过程以减少开销（可认为TCP无需应用层同步）
- 更新了 Receiver, mock_device, InferenceEngine 和 IPC，改用 float16 代替 float32，以提高内存利用率并加快处理速度。

+++ Software/inference_engine.py
@@ -137,13 +137,13 @@ class InferenceEngine:
         # 分配 pinned memory 的 torch tensor，再通过 .numpy() 暴露 numpy 视图，两者共享底层内存
         self.gpu_ring = torch.empty(
             (self.waterfall_height, self.waterfall_width),
-            dtype=torch.float32,
+            dtype=torch.float16,
             device=self.device,
         )
         # 用于 CPU→GPU 的 pinned 中转缓冲（存未排序的原始 ring 数据）
         self._ring_pin = torch.empty(
             (self.waterfall_height, self.waterfall_width),
-            dtype=torch.float32,
+            dtype=torch.float16,
         ).pin_memory()
         self._ring_pin_np = self._ring_pin.numpy()
 
@@ -222,7 +222,7 @@ class InferenceEngine:
                 if self.waterfall_height != self.waterfall_view.shape[0]:
                     self._waterfall_pin = torch.empty(
                         (self.waterfall_height, self.waterfall_width),
-                        dtype=torch.float32,
+                        dtype=torch.float16,
                     ).pin_memory()
                     self.waterfall_view = self._waterfall_pin.numpy()
             elif name == "max_batch_windows":
